+++ src/components/Hero/heroChangelogData.ts
@@ -10,8 +10,8 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
-    date: "Prerelease",
-    version: "Prerelease",
+    date: "2026-02-26",
+    version: "4.1.1",
     type: "Update",
     affects: ["Functionality", "Styles"],
     notes: [